+++ src/init.c
@@ -34,6 +34,7 @@ static void RegisterSymbols(void) {
 static const R_CallMethodDef CallEntries[] = {
   {"rnng_aio_call", (DL_FUNC) &rnng_aio_call, 1},
   {"rnng_aio_get_msg", (DL_FUNC) &rnng_aio_get_msg, 1},
+  {"rnng_aio_http", (DL_FUNC) &rnng_aio_http, 1},
   {"rnng_aio_result", (DL_FUNC) &rnng_aio_result, 1},
   {"rnng_aio_stop", (DL_FUNC) &rnng_aio_stop, 1},
   {"rnng_aio_stream_recv", (DL_FUNC) &rnng_aio_stream_recv, 1},
@@ -72,6 +73,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"rnng_listener_set_uint64", (DL_FUNC) &rnng_listener_set_uint64, 3},
   {"rnng_listener_start", (DL_FUNC) &rnng_listener_start, 1},
   {"rnng_ncurl", (DL_FUNC) &rnng_ncurl, 5},
+  {"rnng_ncurl_aio", (DL_FUNC) &rnng_ncurl_aio, 5},
   {"rnng_protocol_open", (DL_FUNC) &rnng_protocol_open, 1},
   {"rnng_recv", (DL_FUNC) &rnng_recv, 2},
   {"rnng_recv_aio", (DL_FUNC) &rnng_recv_aio, 2},